pyproject.toml
[project]
name = "readers-publication-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Mostafijur Rahman",email = "mostafijur.rahman299@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-unfold (>=0.59.0,<0.60.0)",
    "django-phonenumber-field (>=8.1.0,<9.0.0)",
    "phonenumbers (>=9.0.6,<10.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "twilio (>=9.6.2,<10.0.0)",
    "django-ckeditor-5 (>=0.2.18,<0.3.0)",
    "faker (>=37.4.0,<38.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
